+++ proj/src/mvc/model/game.h
@@ -17,11 +17,11 @@ struct mousePosition{
 };
 
 struct Clock {
-  uint8_t days;
-  uint8_t seconds;
-  uint8_t minutes;
-  uint8_t hours;
-  uint8_t a_tenth_of_a_second;
+  int days;
+  int seconds;
+  int minutes;
+  int hours;
+  int a_tenth_of_a_second;
 };
 
 struct cursor {
@@ -80,17 +80,15 @@ void destroy_game(struct Game* game);
 
 struct Player* create_player();
 void destroy_player(struct Player* player);
-void init_player(struct Player *player, bool isWhite , uint8_t minutes, uint8_t seconds);
+void init_player(struct Player *player, bool isWhite , int minutes, int seconds);
 
-void initClock(struct Clock *clock , uint8_t minutes, uint8_t seconds);
+void initClock(struct Clock *clock , int minutes, int seconds);
 
 void changeState(struct Game *game, enum GameStates state);
 void changeTurn(struct Game *game);
 
 void move_piece(struct Game *game, enum PieceType PieceType, struct Piece * piece,
                 struct Position *init_pos, struct Position *final_pos);
-
-
 
 struct Piece* is_check(struct Game *game);
 bool is_checkmate(struct Game *game);
